requirements.txt
requests
fastapi
fastapi_mcp
wikipedia
google-adk
